requirements.txt
# Core dependencies
ccxt>=4.0.0
torch>=2.0.0
pandas>=2.0.0
matplotlib>=3.7.0
requests>=2.31.0
numpy>=1.24.0
asyncio>=3.4.3

# AI and forecasting
chronos-forecasting>=0.1.0
llama-index>=0.9.0
llama-index-core>=0.9.0
llama-index-embeddings-huggingface>=0.1.0
sentence-transformers>=2.2.0

# Docker and environment
python-dotenv>=1.0.0